% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{gcm}
\alias{gcm}
\title{GCM using random forests}
\usage{
gcm(Y, X, Z, alternative = c("two.sided", "less", "greater"), ...)
}
\arguments{
\item{Y}{Response}

\item{X}{Covariates}

\item{Z}{Covariates}

\item{alternative}{Alternative}

\item{...}{Additional arguments to ranger}
}
\value{
Object of class htest
}
\description{
GCM using random forests
}
\examples{
X <- matrix(rnorm(2e3), ncol = 2)
colnames(X) <- c("X1", "X2")
Z <- matrix(rnorm(2e3), ncol = 2)
colnames(Z) <- c("Z1", "Z2")
Y <- rnorm(1e3) # X[, 2] + Z[, 2] + rnorm(1e3)
(gcm1 <- gcm(Y, X, Z))
plot(gcm1)

}
