% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm.R
\name{pcm}
\alias{pcm}
\title{Conditional mean independence test}
\usage{
pcm(
  Y,
  X,
  Z,
  rep = 1,
  est_vhat = TRUE,
  reg = c("pcm_ranger", "pcm_lasso"),
  ghat_args = NULL,
  mtry = identity,
  do.check = FALSE,
  ...
)
}
\arguments{
\item{Y}{Numeric; response}

\item{X}{Numeric; covariates}

\item{Z}{Numceric; covariates}

\item{rep}{Number of repetitions}

\item{est_vhat}{Estimate variance functional}

\item{reg}{Character; regression method}

\item{ghat_args}{Arguments passed to reg}

\item{mtry}{Argument passed to \code{ranger}}

\item{do.check}{Save check data}

\item{...}{Additional arguments passed to \code{reg}}
}
\value{
Object of class '\code{htest}'
}
\description{
Conditional mean independence test
}
\examples{
X <- matrix(rnorm(2e3), ncol = 2)
colnames(X) <- c("X1", "X2")
Z <- matrix(rnorm(2e3), ncol = 2)
colnames(Z) <- c("Z1", "Z2")
Y <- rnorm(1e3) # X[, 2] + Z[, 2] + rnorm(1e3)
(pcm1 <- pcm(Y, X, Z, ghat_args = list(mtry = NULL, max.depth = NULL),
    est_vhat = TRUE, do.check = TRUE))
plot(pcm1)

}
