% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R, R/pcm.R, R/wgcm.R
\name{plot.gcm}
\alias{plot.gcm}
\alias{plot.pcm}
\alias{plot.wgcm}
\title{Plotting methods for COMETs}
\usage{
\method{plot}{gcm}(x, plot = TRUE, ...)

\method{plot}{pcm}(x, plot = TRUE, ...)

\method{plot}{wgcm}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{Object of class '\code{gcm}', '\code{pcm}', or '\code{wgcm}'.}

\item{plot}{Logical; whether to print the plot (default: \code{TRUE}).}

\item{...}{Currently ignored.}
}
\description{
Plotting methods for COMETs
}
