% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcm.R
\name{wgcm}
\alias{wgcm}
\title{Weighted Generalised covariance measure test}
\usage{
wgcm(
  Y,
  X,
  Z,
  alternative = c("two.sided", "less", "greater"),
  reg_YonZ = "rf",
  reg_XonZ = "rf",
  reg_wfun = "rf",
  args_XonZ = NULL,
  args_wfun = NULL,
  ...
)
}
\arguments{
\item{Y}{Vector of response values. Can be supplied as a numeric vector or
a single column matrix.}

\item{X}{Matrix or data.frame of covariates.}

\item{Z}{Matrix or data.frame of covariates.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}}

\item{reg_YonZ}{Character string or function specifying the regression for
Y on Z.}

\item{reg_XonZ}{Character string or function specifying the regression for
X on Z.}

\item{reg_wfun}{Character string or function specifying the regression for
estimating the weighting function.}

\item{args_XonZ}{Additional arguments passed to \code{reg_XonZ}.}

\item{args_wfun}{Additional arguments passed to \code{reg_XonZ}.}

\item{...}{Additional arguments passed to \code{reg_YonZ}}
}
\value{
Object of class '\code{wgcm}' and '\code{htest}' with the following
components:
\item{\code{statistic}}{The value of the test statistic.}
\item{\code{p.value}}{The p-value for the \code{hypothesis}}
\item{\code{parameter}}{In case X is multidimensional, this is the degrees of
    freedom used for the chi-squared test.}
\item{\code{hypothesis}}{String specifying the null hypothesis .}
\item{\code{null.value}}{String specifying the null hypothesis.}
\item{\code{method}}{The string \code{"Generalised covariance measure test"}.}
\item{\code{data.name}}{A character string giving the name(s) of the data.}
\item{\code{rY}}{Residuals for the Y on Z regression.}
\item{\code{rX}}{Weighted residuals for the X on Z regression.}
}
\description{
Weighted Generalised covariance measure test
}
\details{
The weighted generalised covariance measure test tests whether a weighted
version of the conditional covariance of Y and X given Z is zero.
}
\examples{
X <- matrix(rnorm(3e2), ncol = 2)
colnames(X) <- c("X1", "X2")
Z <- matrix(rnorm(3e2), ncol = 2)
colnames(Z) <- c("Z1", "Z2")
Y <- X[, 2]^2 + Z[, 2] + rnorm(150)
(wgcm1 <- wgcm(Y, X, Z))

}
\references{
Scheidegger, C., Hörrmann, J., & Bühlmann, P. (2022). The weighted
generalised covariance measure. Journal of Machine Learning Research,
23(273), 1-68.
}
